refactor(engine): split HTTP client modules
